+++ src/common.rs
@@ -4,6 +4,7 @@ mod line_entry;
 pub(crate) mod output;
 mod warning;
 
+use colored::*;
 pub use file_entry::FileEntry;
 pub use line_entry::LineEntry;
 pub use output::Output;